tag: Remove global re-exports of tag items
