+++ src/tag/mod.rs
@@ -1,19 +1,26 @@
+//! Utilities for generic tag handling
+
 pub(crate) mod item;
+mod tag_type;
+mod tagext;
 pub(crate) mod utils;
 
 use crate::config::WriteOptions;
 use crate::error::{LoftyError, Result};
-use crate::file::FileType;
 use crate::macros::err;
 use crate::picture::{Picture, PictureType};
 use crate::probe::Probe;
-use crate::traits::{Accessor, MergeTag, SplitTag, TagExt};
-use item::{ItemKey, ItemValue, TagItem};
+use crate::traits::{Accessor, MergeTag, SplitTag};
 
 use std::borrow::Cow;
-use std::fs::{File, OpenOptions};
+use std::fs::File;
 use std::io::Write;
 use std::path::Path;
+
+// Exports
+pub use item::{ItemKey, ItemValue, TagItem};
+pub use tag_type::TagType;
+pub use tagext::TagExt;
 
 macro_rules! impl_accessor {
 	($($item_key:ident => $name:tt),+) => {
@@ -581,6 +588,7 @@ impl TagExt for Tag {
 }
 
 #[derive(Debug, Clone, Default)]
+#[allow(missing_docs)]
 pub struct SplitTagRemainder;
 
 impl SplitTag for Tag {
@@ -599,74 +607,14 @@ impl MergeTag for SplitTagRemainder {
 	}
 }
 
-/// The tag's format
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum TagType {
-	/// This covers both APEv1 and APEv2 as it doesn't matter much
-	Ape,
-	/// Represents an ID3v1 tag
-	Id3v1,
-	/// This covers all ID3v2 versions since they all get upgraded to ID3v2.4
-	Id3v2,
-	/// Represents an MP4 ilst atom
-	Mp4Ilst,
-	/// Represents vorbis comments
-	VorbisComments,
-	/// Represents a RIFF INFO LIST
-	RiffInfo,
-	/// Represents AIFF text chunks
-	AiffText,
-}
-
-impl TagType {
-	/// Remove a tag from a [`Path`]
-	///
-	/// # Errors
-	///
-	/// See [`TagType::remove_from`]
-	pub fn remove_from_path(&self, path: impl AsRef<Path>) -> Result<()> {
-		let mut file = OpenOptions::new().read(true).write(true).open(path)?;
-		self.remove_from(&mut file)
-	}
-
-	#[allow(clippy::shadow_unrelated)]
-	/// Remove a tag from a [`File`]
-	///
-	/// # Errors
-	///
-	/// * It is unable to guess the file format
-	/// * The format doesn't support the tag
-	/// * It is unable to write to the file
-	pub fn remove_from(&self, file: &mut File) -> Result<()> {
-		let probe = Probe::new(file).guess_file_type()?;
-		let Some(file_type) = probe.file_type() else {
-			err!(UnknownFormat);
-		};
-
-		// TODO: This should not have to be manually updated
-		let special_exceptions = ((file_type == FileType::Ape
-			|| file_type == FileType::Mpc
-			|| file_type == FileType::Flac)
-			&& *self == TagType::Id3v2)
-			|| file_type == FileType::Mpc && *self == TagType::Id3v1;
-
-		if !special_exceptions && !file_type.supports_tag_type(*self) {
-			err!(UnsupportedTag);
-		}
-
-		let file = probe.into_inner();
-		utils::write_tag(&Tag::new(*self), file, file_type, WriteOptions::default()) // TODO
-	}
-}
-
 #[cfg(test)]
 mod tests {
 	use super::try_parse_year;
 	use crate::config::WriteOptions;
 	use crate::prelude::*;
 	use crate::tag::utils::test_utils::read_path;
-	use crate::{Picture, PictureType, Tag, TagType};
+	use crate::tag::{Tag, TagType};
+	use crate::{Picture, PictureType};
 
 	use std::io::{Seek, Write};
 	use std::process::Command;
