+++ tests/files/aac.rs
@@ -2,7 +2,8 @@ use crate::{set_artist, temp_file, verify_artist};
 use lofty::config::ParseOptions;
 use lofty::file::FileType;
 use lofty::prelude::*;
-use lofty::{Probe, TagType};
+use lofty::tag::TagType;
+use lofty::Probe;
 
 use std::io::{Seek, Write};
